Update packages for OpenWrt 24.10 and switch CI over to the new version (#31)
Due to changes to package downloads from git, the PKG_MIRROR_HASH of those
packages will only match when building on OpenWrt 24.10; this limits the
ability for these package Makefiles to be backwards compatible with OpenWrt
23.05 without additional effort. This change is therefore dropping
compatibility with that release, and also removing workarounds for other issues
that are fixed in OpenWrt 24.10.

Split openthread-br-luci into a separate package, since it requires
luci-lua-runtime which is no longer installed by default in 24.10.

Ensure build container and .config.ci have consistent CONFIG_ALL and
CONFIG_AUTOREMOVE settings.